[fix] OS does not start in ro-root mode and makes fsck per default
The default behaviour should be that the system starts up
with a read-only root FS and no FS check is done (faster boot).
To make easier to remember the two files that control the
startup are renamed /do-not-disable-root-ro and /do-not-forcefsck.
If you need to start in rw mode or want a fsck run, rename the files
by removing "do-not".